refactor(075): extract @fairwins/intent-types — one EIP-712 source (US4, extraction only)
STRUCTURAL ONLY. No behaviour change: this commit MOVES the struct definitions
and changes nothing about what any action does. The functional fix (the missing
gateway action) is the next commit, deliberately separate so a bisect can tell
"moved" from "added".

WHAT WAS WRONG
These field lists decide what a member's signature AUTHORISES. A mismatch
between signer and verifier does not raise an error — it produces a signature
that verifies against something other than what the member was shown. CLAUDE.md
required three copies to stay byte-identical (contract typehashes, frontend,
relay gateway) and enforced it with human discipline.

Discipline held for 26 of 27 structs. The 27th did not: `InvalidateNonce`
existed in SignerIntentBase.sol and in the frontend and was ABSENT from the
gateway. There was also a FOURTH copy nobody counted — the EIP-3009
ReceiveWithAuthorization table, duplicated in frontend/src/lib/pools/gasless.js
and in the gateway. That one is the money path.

WHY IT COULDN'T JUST BE IMPORTED BEFORE
frontend/src uses extensionless relative imports (Vite-resolved) — 2,966 of
them — while the gateway is plain Node ESM, which requires extensions. The
gateway physically could not import the frontend's copy, so it kept its own.
The new package is authored Node-resolvable (extensioned imports + explicit
`exports`), which is what makes one source possible at all.

  packages/intent-types  27 structs + EIP-3009 + typeStringFor(); zero runtime deps
  frontend intentTypes   329 -> 165 lines (keeps INTENT_ACTIONS + domain helpers)
  gateway  intentTypes   622 -> 480 lines (keeps its ethers calldata machinery)
  pools/gasless.js       local EIP-3009 literal -> re-export

THE GATE THAT REPLACES THE CONVENTION
test/intent/TypehashParity.test.js parses `keccak256("<literal>")` out of the
production contracts and asserts our rendered type string is identical, plus
the keccak equality. Strings rather than hashes because when it fails you see
WHICH FIELD moved, not that two opaque 32-byte values differ. contracts/mocks/
is excluded: a mock's typehash proves nothing about what a deployed contract
verifies.

All 27 structs match their verifying contract exactly, so the extraction is
faithful.

EIP-3009 IS CHECKED DIFFERENTLY, ON PURPOSE
Its authoritative typehash lives in DEPLOYED USDC — Circle's contract, not
ours. The only in-repo Solidity copy is MockUSDCPermit.sol, which we wrote, so
checking against it would be circular: both could drift together and the gate
would stay green while every gasless payment stopped verifying. It is pinned to
a fixed vector instead, which I verified by reading
RECEIVE_WITH_AUTHORIZATION_TYPEHASH live off USDC on Polygon 137
(0x3c499c54…) — 0xd099cc98… — so the vector is independent, not a hash of our
own string.

VERIFIED: TypehashParity 6 passing; gateway 226 passing; frontend intent/relay/
pool suites 60 passing across 7 files.
